refactor testcluster out of library as its used in specs only
