+++ spec/spec_helper.rb
@@ -21,4 +21,25 @@
 
 require_relative '../lib/crate_ruby'
 require 'net/http'
+require_relative 'support/test_cluster'
+
+HOST = '127.0.0.1'
+PORT = 44200
+
+RSpec.configure do |config|
+  config.before(:each) do
+  end
+  config.after(:each) do
+  end
+  config.before(:suite) do
+    @cluster = TestCluster.new(1, PORT)
+    @cluster.start_nodes
+  end
+  config.after(:suite) do
+    pid_file = File.join(__dir__, 'support/testnode.pid')
+    pid = File.read(pid_file)
+    File.delete(pid_file)
+    Process.kill('HUP', pid.to_i)
+  end
+end
